requirements.txt
requests==2.31.0
openai==1.3.0
beautifulsoup4==4.12.3
python-telegram-bot==20.7
